% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.r
\name{coerce}
\alias{coerce}
\title{Coerce an object into a particular type, with handling for NULL.}
\usage{
coerce(obj, type)
}
\arguments{
\item{obj}{The object to coerce.}

\item{type}{The type to coerce it to.}
}
\description{
Coerce an object into a particular type, with handling for NULL.
}

