\name{extract_params_from_request}
\alias{extract_params_from_request}
\title{Helper function for extracting a list of POST parameters from a request
body (in JSON format for now).}
\usage{
extract_params_from_request(request)
}
\arguments{
  \item{request}{environment. The httpuv request
  environment.}
}
\description{
Helper function for extracting a list of POST parameters
from a request body (in JSON format for now).
}

