\name{extract_query_from_request}
\alias{extract_query_from_request}
\title{Helper function for extracting a list of GET parameters from a request
body (in JSON format for now).}
\usage{
extract_query_from_request(request)
}
\arguments{
  \item{request}{environment. The httpuv request
  environment.}
}
\description{
Helper function for extracting a list of GET parameters
from a request body (in JSON format for now).
}

