% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/html.r
\name{fetch_asset}
\alias{fetch_asset}
\title{Legacy fetch_asset code.}
\usage{
fetch_asset(asset_path)
}
\arguments{
\item{asset_path}{character. File path to the asset to be hosted.}
}
\description{
Legacy fetch_asset code.
}

