% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/html.r
\name{html}
\alias{html}
\title{Returns HTML in a server.}
\usage{
html(html_string)
}
\arguments{
\item{html_string}{character. A sring of HTML.}
}
\description{
Returns HTML in a server.
}
\examples{
\dontrun{
   routes <- list("/index" = function(...) html("<b>bold!</b>"))
}
}

