% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/html.r
\name{html_page}
\alias{html_page}
\title{Parses an HTML page and returns it to the server.}
\usage{
html_page(file)
}
\arguments{
\item{file}{character. The relative path of the file.}
}
\description{
Parses an HTML page and returns it to the server.
}
\examples{
\dontrun{
   routes <- list("/index" = html_page("views/index.html")
}
}

