% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/response.r
\name{is.microserver_response}
\alias{is.microserver_response}
\title{Determine whether something is of S3 class microserver_response.}
\usage{
is.microserver_response(obj)
}
\arguments{
\item{obj}{any R object.}
}
\value{
\code{TRUE} or \code{FALSE} if \code{obj} inherits \code{response}.
}
\description{
Determine whether something is of S3 class microserver_response.
}
\seealso{
\code{\link{response}}
}

