% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.r
\name{is_simple_list}
\alias{is_simple_list}
\title{Check if an object is a simple list (i.e., it is a list, but not a list of lists).}
\usage{
is_simple_list(lst)
}
\arguments{
\item{lst}{list. A list to check.}
}
\description{
Check if an object is a simple list (i.e., it is a list, but not a list of lists).
}

