\docType{package}
\name{microserver}
\alias{microserver}
\alias{microserver-package}
\title{Microserver is a package that builds on the httupv library
and aims to provide a minimal set of tools to get a JSON-based
 REST API off the ground.}
\description{
Microserver is a package that builds on the httupv library
and aims to provide a minimal set of tools to get a
JSON-based REST API off the ground.
}

