% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/response.r
\name{microserver_response}
\alias{microserver_response}
\title{Simple S3 class to denote JSON responses for httupv.}
\usage{
microserver_response(response = NULL, status = 200,
  headers = list(`content-type` = "text/json"))
}
\arguments{
\item{response}{ANY. The R object to send as a response parameter.
If the header is set to text/json it will be converted into a JSON string.}

\item{status}{integer. HTTP status (default is \code{200}).}

\item{headers}{list. A list of HTTP headers (default is
\code{list('Content-Type' = 'text/json')}.}
}
\description{
Simple S3 class to denote JSON responses for httupv.
}
\examples{
\dontrun{
response('Did not work!', 404) # 404 error
response(list(a = 1, b = 2))   # { "a": 1, "b": 2 } JSON response
}
}

