% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.r
\name{recursively_enlist_if_named}
\alias{recursively_enlist_if_named}
\title{Recursively run the enlist function.}
\usage{
recursively_enlist_if_named(obj)
}
\arguments{
\item{obj}{An object to enlist.}
}
\description{
Recursively run the enlist function.
}

