% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{simplify_homogeneous_lists}
\alias{simplify_homogeneous_lists}
\title{Fix jsonlite's JSON simplification.}
\usage{
simplify_homogeneous_lists(object)
}
\arguments{
\item{object}{any R object derived from \code{\link[jsonlite]{fromJSON}}.}
}
\value{
the same object, with any list components where each element is
  an atomic vector of length 1 or NULL coerced to a single atomic vector.
}
\description{
Fix jsonlite's JSON simplification.
}
\note{
See https://github.com/jeroenooms/jsonlite/issues/66 for more details.
}
\examples{
\dontrun{
  simplify_homogeneous_lists(jsonlite::fromJSON(
   '{ "numeric": [1,2], "list": [1, "a"] }', simplifyVector = FALSE))
  # A list with atomic numeric vector in the "numeric" key and
  # a list in the "list" key.
  # list(numeric = c(1,2), list = list(1, "a"))
}
}

