% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.r
\name{terminal}
\alias{terminal}
\title{TRUE or FALSE if the object is a terminal.
A terminal is either a NULL or a length-1 atomic.}
\usage{
terminal(x)
}
\description{
TRUE or FALSE if the object is a terminal.
A terminal is either a NULL or a length-1 atomic.
}

