% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.r
\name{enlist_if_named}
\alias{enlist_if_named}
\title{Convert an object to a list, but only if that object has names.}
\usage{
enlist_if_named(obj)
}
\description{
Convert an object to a list, but only if that object has names.
}

