% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/request.r
\name{extract_params_from_request}
\alias{extract_params_from_request}
\title{Helper function for extracting a list of POST parameters from a request
body (in JSON format for now).}
\usage{
extract_params_from_request(request)
}
\arguments{
\item{request}{environment. The httpuv request environment.}
}
\value{
a list with the processed POST parameters.
}
\description{
Helper function for extracting a list of POST parameters from a request
body (in JSON format for now).
}

