% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/request.r
\name{extract_query_from_request}
\alias{extract_query_from_request}
\title{Helper function for extracting a list of GET parameters from a request
body (in JSON format for now).}
\usage{
extract_query_from_request(request)
}
\arguments{
\item{request}{environment. The httpuv request environment.}
}
\value{
a list with the processed GET parameters.
}
\description{
Helper function for extracting a list of GET parameters from a request
body (in JSON format for now).
}

