% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/microserver-package.r
\docType{package}
\name{microserver}
\alias{microserver}
\alias{microserver-package}
\title{Microserver is a package that builds on the httupv library
and aims to provide a minimal set of tools to get a JSON-based
REST API off the ground.}
\description{
Microserver is a package that builds on the httupv library
and aims to provide a minimal set of tools to get a JSON-based
REST API off the ground.
}

