% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/microserver.r
\name{run_server}
\alias{run_server}
\title{Minimal function for opening a socket and accepting/responding to
requests.}
\usage{
run_server(routes, port = 8103)
}
\arguments{
\item{port}{integer. The default is 8103.}
}
\description{
Minimal function for opening a socket and accepting/responding to
requests.
}

