% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microserver.r
\name{run_server}
\alias{run_server}
\title{Minimal function for opening a socket and accepting/responding to
requests.}
\usage{
run_server(routes, port = 8103, http_basic_auth = NULL)
}
\arguments{
\item{routes}{list. A named list of routes.}

\item{port}{integer. The default is 8103.}

\item{http_basic_auth}{list. A list of HTTP basic authentication requirements.
If not \code{NULL} (the default), should contain a \code{token} and
\code{routes} key.}
}
\description{
Minimal function for opening a socket and accepting/responding to
requests.
}

